+++ sphinx/conf.py
@@ -21,7 +21,7 @@ copyright = '2021, LEA, Paderborn University'
 author = 'LEA-UPB'
 
 # The full version, including alpha/beta/rc tags
-release = '0.4.0'
+release = '0.4.1'
 
 # -- General configuration ---------------------------------------------------