+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - #93: GUI Manufacturer parameters for switch and diode must be optional
 - Fix GUI measurement integration method when integrating own measurement.
 - Fix some GUI data does not accept float or negative values
+- Fix some wrong labels: Units in C-curves
 
 ## [0.4.0] - 2022-06-20
 ### Added